refactor: update folder name
