+++ class-components/src/App.css
@@ -13,28 +13,9 @@
   background-repeat: no-repeat;
   background-size: 50%;
 }
-.top-section {
-  max-width: 80%;
-  margin: 0 auto;
-  border: 2px solid black;
-  height: 50px;
-  display: flex;
-  align-items: center;
-  border-radius: 5px;
-}
-.top-section div {
-  padding-left: 20px;
-}
 
-.bottom-section{
-  max-width: 80%;
-  margin: 0 auto;
-}
 
-.item {
- padding: 10px 0;
- border-bottom: 2px solid rgb(209, 209, 209);
-}
+
 
 .btn {
     width: 100px;
