+++ class-components/src/components/BottomSection/BottomSection.css
@@ -1,0 +1,9 @@
+.bottom-section {
+    max-width: 80%;
+    margin: 0 auto;
+  }
+
+.item {
+   padding: 10px 0;
+   border-bottom: 2px solid rgb(209, 209, 209);
+  }
