+++ class-components/src/components/ErrorBoundary/ErrorBoundary.type.ts
@@ -1,0 +1,8 @@
+import { ReactNode } from 'react';
+
+export interface ErrorBoundaryProps {
+  children: ReactNode;
+}
+export interface ErrorBoundaryState {
+  hasError: boolean;
+}
