+++ class-components/src/components/Search/Search.type.ts
@@ -1,0 +1,7 @@
+export interface SearchState {
+  searchTerm: string;
+}
+
+export interface SearchProps {
+  onSearch: (searchTerm: string) => void;
+}
