+++ class-components/src/components/TopSection/TopSection.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
-import Search from './Search';
-import { SearchProps } from '../types/interfaces';
+import Search from '../Search/Search';
+import { SearchProps } from '../../scripts/types/interfaces';
+import './TopSection.css';
 
 const TopSection: React.FC<SearchProps> = ({ onSearch }) => {
   return (
