+++ class-components/src/components/TopSection/TopSection.type.ts
@@ -1,0 +1,3 @@
+export interface SearchProps {
+  onSearch: (searchTerm: string) => void;
+}
